running main through python
